+++ lib/src/widgets/spaced_column.dart
@@ -9,7 +9,31 @@ import 'package:fp_util/fp_util.dart';
 class SpacedColumn extends Column {
   SpacedColumn({
     super.key,
-    double spacing = 0,
+    double spacing = 10,
+    bool hasLeadingSpace = false,
+    bool hasTrailingSpace = false,
+    super.mainAxisAlignment,
+    super.mainAxisSize,
+    super.crossAxisAlignment,
+    super.textBaseline,
+    super.textDirection,
+    super.verticalDirection,
+    List<Widget> children = const [],
+    bool responsive = false,
+  }) : super(children: [
+          if (hasLeadingSpace) spacing.verticalSpace(responsive),
+          ...children.expandIndexed((index, widget) {
+            return [
+              widget,
+              if (index < children.length || hasTrailingSpace)
+                spacing.verticalSpace(responsive),
+            ];
+          }),
+        ]);
+
+  SpacedColumn.separated({
+    super.key,
+    Widget separator = const SizedBox(height: 10),
     bool hasLeadingSpace = false,
     bool hasTrailingSpace = false,
     super.mainAxisAlignment,
@@ -20,12 +44,11 @@ class SpacedColumn extends Column {
     super.verticalDirection,
     List<Widget> children = const [],
   }) : super(children: [
-          if (hasLeadingSpace) spacing.verticalSpace,
+          if (hasLeadingSpace) separator,
           ...children.expandIndexed((index, widget) {
             return [
               widget,
-              if (index < children.length || hasTrailingSpace)
-                spacing.verticalSpace,
+              if (index < children.length || hasTrailingSpace) separator,
             ];
           }),
         ]);