+++ lib/src/widgets/spaced_row.dart
@@ -9,7 +9,32 @@ import 'package:fp_util/src/extensions/extensions.dart';
 class SpacedRow extends Row {
   SpacedRow({
     super.key,
-    double spacing = 8.0,
+    double spacing = 10,
+    bool hasLeadingSpace = false,
+    bool hasTrailingSpace = false,
+    super.mainAxisAlignment,
+    super.crossAxisAlignment,
+    super.mainAxisSize,
+    super.textBaseline,
+    super.textDirection,
+    super.verticalDirection,
+    List<Widget> children = const [],
+    bool responsive = false,
+  }) : super(children: [
+          if (hasLeadingSpace) spacing.horizontalSpace(responsive),
+          ...children.expandIndexed(
+            (index, child) {
+              return [
+                child,
+                if (index < children.length || hasTrailingSpace)
+                  spacing.horizontalSpace(responsive),
+              ];
+            },
+          ),
+        ]);
+  SpacedRow.separated({
+    super.key,
+    Widget separator = const SizedBox(width: 10),
     bool hasLeadingSpace = false,
     bool hasTrailingSpace = false,
     super.mainAxisAlignment,
@@ -20,13 +45,12 @@ class SpacedRow extends Row {
     super.verticalDirection,
     List<Widget> children = const [],
   }) : super(children: [
-          if (hasLeadingSpace) spacing.horizontalSpace,
+          if (hasLeadingSpace) separator,
           ...children.expandIndexed(
             (index, child) {
               return [
                 child,
-                if (index < children.length || hasTrailingSpace)
-                  spacing.horizontalSpace,
+                if (index < children.length || hasTrailingSpace) separator,
               ];
             },
           ),