+++ lib/src/widgets/widgets.dart
@@ -11,6 +11,8 @@ export 'padded_column.dart';
 export 'padded_row.dart';
 export 'padded_spaced_column.dart';
 export 'padded_spaced_row.dart';
+export 'responsive_padding.dart';
+export 'responsive_sized_box.dart';
 export 'scaffold_wrapper.dart';
 export 'scrollable_column.dart';
 export 'scrollable_row.dart';